local reduction in pmapsum_commutative
pmapsum_commutative carries out a local reduction
on each host, followed by a final reduction across hosts.
This is achieved by using a SegmentedSequentialBinaryTree,
that is a collection of SequentialBinaryTrees on each host,
with an extra tree that connects the hosts where the eventual
reduction is carred out.

This commit also changes procs_node and nprocs_node to return
an OrderedDict instead of a Dict, in order to preserve the sequence
of hosts.